src/offline/requirements.txt
# ── offline/requirements.txt ──────────────────────────────────────────────────
# Install in Colab with: pip install -r requirements.txt
transformers==4.41.0
torch==2.3.0
Pillow==10.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tqdm==4.66.4
matplotlib==3.9.0
requests==2.31.0
psycopg2-binary==2.9.9
